Moved ExtractMethod out of the RefactorMenu into it's own Command , see #761
