+++ RetailCoder.VBE/Rubberduck.csproj
@@ -361,7 +361,6 @@
     <Compile Include="UI\Command\MenuItems\OptionsCommandMenuItem.cs" />
     <Compile Include="UI\Command\MenuItems\ParentMenus\ParentMenuItemBase.cs" />
     <Compile Include="UI\Command\MenuItems\ParentMenus\ProjectWindowContextParentMenu.cs" />
-    <Compile Include="UI\Command\RefactorExtractMethodCommand.cs" />
     <Compile Include="UI\Command\MenuItems\RefactorExtractMethodCommandMenuItem.cs" />
     <Compile Include="UI\Command\MenuItems\ParentMenus\RefactoringsParentMenu.cs" />
     <Compile Include="UI\Command\RefactorRemoveParametersCommand.cs" />
@@ -719,6 +718,8 @@
       <DependentUpon>UnSyncedCommitsControl.cs</DependentUpon>
     </Compile>
     <Compile Include="UnitTesting\PermissiveAssertClass.cs" />
+    <Compile Include="UI\Command\Refactorings\RefactorCommandBase.cs" />
+    <Compile Include="UI\Command\Refactorings\RefactorExtractMethodCommand.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="UI\AboutWindow.resx">
@@ -868,6 +869,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Properties\DataSources\" />
+    <Folder Include="UI\Command\Refactorings\" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\rubberduck.config" />
