--- RetailCoder.VBE/UI/Command/RefactorExtractMethodCommand.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Runtime.InteropServices;
-
-namespace Rubberduck.UI.Command
-{
-    [ComVisible(false)]
-    public class RefactorExtractMethodCommand : CommandBase
-    {
-        public override void Execute(object parameter)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
